+++ src/runway/commands/__init__.py
@@ -5,7 +5,6 @@ from .runway import gen_sample  # noqa
 from .runway import gitclean  # noqa
 from .runway import init # noqa
 from .runway import preflight   # noqa
-from .runway import run_python  # noqa
 from .runway import run_aws  # noqa
 from .runway import run_python  # noqa
 from .runway import run_stacker  # noqa