+++ src/runway/commands/command_loader.py
@@ -20,5 +20,5 @@ def find_command_class(possible_command_names):
 
 def _not_base_class(name_class_pair):
     return name_class_pair[0] not in ['BaseCommand', 'CalledProcessError',
-                                      'RunwayCommand', 'ModulesCommand',
-                                      'Config']
+                                      'Context', 'Dict', 'RunwayCommand',
+                                      'ModulesCommand', 'Config']